New insert and front page single image removed
Attempt to create a new insert for templates "page_navigation_string"
which would be used to display a linked list of pages that are
available instead of just a previous/next option. Also, removed the 1
image on first page limit, all themes will show the current max of 12
images per page on all pages.